requirements.txt
# Core LangChain packages
langchain-community>=0.0.20
langchain-core>=0.1.0
langchain-groq>=0.0.1
langchain-openai>=0.0.5
langgraph>=0.0.20

# Database
pymysql>=1.1.0

# Environment variables
python-dotenv>=1.0.0

# GitHub Inference API
azure-ai-inference>=1.0.0
azure-core>=1.29.0

# For the clustering approach in main.py
pandas>=2.0.0
sentence-transformers>=2.2.0
scikit-learn>=1.3.0

# Additional dependencies for LangChain integration
typing-extensions>=4.0.0

# UI
streamlit>=1.36.0

# Visualization for relationship diagram
graphviz>=0.20.3

# Google Gemini (LangChain integration)
langchain-google-genai>=1.0.0
google-generativeai>=0.7.0

# PPT generation
python-pptx>=0.6.23 